Start of porting the level generator from Minecraft Beta. Also added level save timer.

+++ SMP/Server/World/Generator/ChunkGen.cs
@@ -30,6 +30,6 @@ namespace SMP {
         /// <param name="w"></param>
         /// <param name="c"></param>
         public abstract void Generate( World w, Chunk c );
-        public abstract void SetSeed( int seed );
+        public abstract void SetSeed( long seed );
     }
 }
